+++ apps/vmq_server/src/vmq_server.erl
@@ -50,23 +50,31 @@ start() ->
 
 -spec stop() -> 'ok'.
 stop() ->
-    _ = [application:stop(App) || App <- [vmq_server,
+    application:stop(vmq_server),
+    wait_until_metadata_has_stopped(),
+    _ = [application:stop(App) || App <- [vmq_plugin,
+                                          riak_sysmon,
                                           clique,
-                                          plumtree,
-                                          jobs,
-                                          eleveldb,
-                                          vmq_plumtree,
                                           asn1,
                                           public_key,
-                                          vmq_plugin,
                                           cowboy,
                                           ranch,
                                           crypto,
                                           ssl,
-                                          riak_sysmon,
                                           os_mon,
                                           lager]],
     ok.
+
+wait_until_metadata_has_stopped() ->
+    Impl = application:get_env(vmq_server, metadata_impl, vmq_plumtree),
+    case lists:keymember(Impl, 1, application:which_applications()) of
+        true ->
+            timer:sleep(100),
+            wait_until_metadata_has_stopped();
+        false ->
+            ok
+    end.
+
 
 maybe_start_distribution() ->
     case ets:info(sys_dist) of